Change stored spillPeriod_s from float to double
When it's stored as a float, `t_event` in the hdf5 files gets a
numerical error of O(1e-6), which means that if you do `t_vert -
t_event`, hoping to get the vertex times relative to the start of the
spill, the spill microstructure will get washed out, as Brooke
discovered.

+++ run-spill-build/overlaySinglesIntoSpillsSorted.C
@@ -281,7 +281,7 @@ void overlaySinglesIntoSpillsSorted(std::string inFileName1,
   geom->Write();
   new_tree->Write();
   event_spill_map->Write("event_spill_map", 1);
-  auto p = new TParameter<float>("spillPeriod_s", spillPeriod_s);
+  auto p = new TParameter<double>("spillPeriod_s", spillPeriod_s);
   p->Write();
 
   outFile->mkdir("DetSimPassThru");
